+++ sdk/net/http/http2/hpack.hpp
@@ -231,6 +231,12 @@ class hpack_session : public http_header_compression_session {
    public:
     hpack_session();
     /**
+     * @brief   duplicate
+     * @param   const std::string& name [in]
+     * @param   const std::string& value [in]
+     */
+    virtual return_t duplicate(const std::string& name, const std::string& value);
+    /**
      * @brief   HPACK query function
      * @param   int cmd [in] see header_compression_cmd_t
      * @param   void* req [in]