+++ sdk/net/http/http2/http_header_compression_session.cpp
@@ -8,11 +8,94 @@
  * Date         Name                Description
  */
 
+#include <math.h>
+
 #include <sdk/net/http/http2/http_header_compression.hpp>
 #include <sdk/net/http/http_resource.hpp>
 
 namespace hotplace {
 namespace net {
+
+return_t qpack_ric2eic(size_t capacity, size_t ric, size_t base, size_t& eic, bool& sign, size_t& deltabase) {
+    return_t ret = errorcode_t::success;
+    /* RFC 9204 4.5.1.1.  Required Insert Count
+     *  if (ReqInsertCount) EncInsertCount = (ReqInsertCount mod (2 * MaxEntries)) + 1
+     *  else EncInsertCount = 0;
+     */
+    if (0 == ric) {
+        eic = ric;
+    } else {
+        size_t maxentries = ::floor(capacity / 32);
+        eic = (ric % (2 * maxentries)) + 1;
+    }
+
+    /* RFC 9204 4.5.1.2.  Base
+     *  A Sign bit of 1 indicates that the Base is less than the Required Insert Count
+     *
+     *  if (0 == Sign) DeltaBase = Base - ReqInsertCount
+     *  else DeltaBase = ReqInsertCount - Base - 1
+     */
+    sign = (ric > base);
+    if (ric > base) {
+        deltabase = ric - base - 1;
+    } else {
+        deltabase = ric - base;
+    }
+
+    return ret;
+}
+
+return_t qpack_eic2ric(size_t capacity, size_t tni, size_t eic, bool sign, size_t deltabase, size_t& ric, size_t& base) {
+    return_t ret = errorcode_t::success;
+    __try2 {
+        // RFC 9204 4.5.1.1.  Required Insert Count
+        size_t maxentries = ::floor(capacity / 32);
+        eic = (ric % (2 * maxentries)) + 1;
+        size_t fullrange = 2 * maxentries;
+        if (0 == eic) {
+            ric = 0;
+        } else {
+            if (eic > fullrange) {
+                ret = errorcode_t::invalid_request;
+                __leave2;
+            }
+
+            size_t maxvalue = tni + maxentries;
+            size_t maxwrapped = ::floor(maxvalue / fullrange) * fullrange;
+            ric = maxwrapped + eic - 1;
+
+            if (ric > maxvalue) {
+                if (ric <= fullrange) {
+                    ret = errorcode_t::invalid_request;
+                    __leave2;
+                } else {
+                    ric -= fullrange;
+                }
+            }
+
+            if (0 == ric) {
+                ret = errorcode_t::invalid_request;
+                __leave2;
+            }
+
+            /* RFC 9204 4.5.1.2.  Base
+             *  A Sign bit of 1 indicates that the Base is less than the Required Insert Count
+             *
+             *  if (0 == Sign) Base = DeltaBase + ReqInsertCount
+             *  else Base = ReqInsertCount - DeltaBase - 1
+             */
+            if (0 == sign) {
+                base = deltabase + ric;
+            } else {
+                base = ric - deltabase - 1;
+            }
+        }
+    }
+    __finally2 {
+        // do nothing
+    }
+    return ret;
+}
 
 http_header_compression_session::http_header_compression_session() : _separate(false), _inserted(0), _dropped(0), _capacity(0x10000), _tablesize(0) {}
 
@@ -32,35 +115,45 @@ bool http_header_compression_session::operator!=(const http_header_compression_s
     return (_separate != rhs._separate) || (_dynamic_map != rhs._dynamic_map);
 }
 
+void http_header_compression_session::trace(std::function<void(stream_t*)> f) { _df = f; }
+
 match_result_t http_header_compression_session::match(const std::string& name, const std::string& value, size_t& index) {
     match_result_t state = match_result_t::not_matched;
 
     auto lbound = _dynamic_map.lower_bound(name);
     auto ubound = _dynamic_map.upper_bound(name);
+    std::priority_queue<size_t> pq;
+
     for (auto iter = lbound; iter != ubound; iter++) {
         const auto& k = iter->first;
-        const auto& v = iter->second;
-        if ((name == k) && (value == v.first)) {
-            state = match_result_t::all_matched_dynamic;
-            /**
-             * get index from v.second
-             *
-             * consider following cases
-             *  capacity = 3, _inserted = 2, _dropped = 0, table {1 0}, table.size = 2
-             *  capacity = 3, _inserted = 3, _dropped = 0, table {2 1 0}, table.size = 3
-             *  capacity = 3, _inserted = 4, _dropped = 1, table {3 2 1}, table.size = 3
-             *  capacity = 3, _inserted = 5, _dropped = 2, table {4 3 2}, table.size = 3
-             *
-             * conclusion
-             *  index = _inserted - _dropped - v.second + _dropped - 1 = _inserted - v.second - 1
-             */
-            index = _inserted - v.second - 1;
-            if (false == _separate) {
-                // HPACK
-                auto static_entries = http_resource::get_instance()->sizeof_hpack_static_table_entries();
-                index += (static_entries + 1);
-            }
-            break;
+        const auto& v = iter->second;  // pair(value, entry)
+        const auto& val = v.first;
+        const auto& ent = v.second;
+        if ((name == k) && (value == val)) {
+            pq.push(ent);
+        }
+    }
+
+    if (pq.size()) {
+        state = match_result_t::all_matched_dynamic;
+        const auto& ent = pq.top();
+        /**
+         * get index from v.second
+         *
+         * consider following cases
+         *  capacity = 3, _inserted = 2, _dropped = 0, table {1 0}, table.size = 2
+         *  capacity = 3, _inserted = 3, _dropped = 0, table {2 1 0}, table.size = 3
+         *  capacity = 3, _inserted = 4, _dropped = 1, table {3 2 1}, table.size = 3
+         *  capacity = 3, _inserted = 5, _dropped = 2, table {4 3 2}, table.size = 3
+         *
+         * conclusion
+         *  index = _inserted - _dropped - v.second + _dropped - 1 = _inserted - v.second - 1
+         */
+        index = _inserted - ent - 1;
+        if (false == _separate) {
+            // HPACK
+            auto static_entries = http_resource::get_instance()->sizeof_hpack_static_table_entries();
+            index += (static_entries + 1);
         }
     }
 
@@ -90,15 +183,18 @@ return_t http_header_compression_session::select(size_t index, std::string& name
             const auto& t = _inserted - index - 1;
             auto riter = _dynamic_reversemap.find(t);
             // never happen (_dynamic_reversemap.end() == riter)
-            const auto& k = riter->second;
-            auto lbound = _dynamic_map.lower_bound(k);
-            auto ubound = _dynamic_map.upper_bound(k);
+            const auto& pne = riter->second;  // pair(name, entry size)
+            const auto& nam = pne.first;
+            auto lbound = _dynamic_map.lower_bound(nam);
+            auto ubound = _dynamic_map.upper_bound(nam);
 
             for (auto iter = lbound; iter != ubound; iter++) {
-                const auto& v = iter->second;
-                if (t == v.second) {
-                    name = k;
-                    value = v.first;
+                const auto& pve = iter->second;  // pair(value, entry)
+                const auto& val = pve.first;
+                const auto& ent = pve.second;
+                if (t == ent) {
+                    name = nam;
+                    value = val;
                     break;
                 }
             }
@@ -136,8 +232,7 @@ return_t http_header_compression_session::insert(const std::string& name, const 
     evict();
 
     _dynamic_map.insert({name, {value, _inserted}});
-    _dynamic_reversemap.insert({_inserted, name});
-    _entry_size.insert({_inserted, entrysize});
+    _dynamic_reversemap.insert({_inserted, {name, entrysize}});
 
     _inserted++;
 
@@ -151,21 +246,30 @@ return_t http_header_compression_session::evict() {
         // RFC 7541 4.2.  Maximum Table Size
         // RFC 7541 4.4.  Entry Eviction When Adding New Entries
         // RFC 9204 3.2.2.  Dynamic Table Capacity and Eviction
-        auto back = _dynamic_reversemap.find(_dropped);
+        auto entry = _dropped;
+        auto back = _dynamic_reversemap.find(entry);
 
-        auto const& t = back->first;
-        auto const& k = back->second;
+        auto const& t = back->first;   // entry
+        auto const& k = back->second;  // (name, entry size)
 
-        auto dpiter = _entry_size.find(t);
-        _tablesize -= dpiter->second;
-        _entry_size.erase(dpiter);
+        auto const& name = k.first;
+        auto const& entrysize = k.second;
 
-        auto lbound = _dynamic_map.lower_bound(k);
-        auto ubound = _dynamic_map.upper_bound(k);
+        _tablesize -= entrysize;
+
+        auto lbound = _dynamic_map.lower_bound(name);
+        auto ubound = _dynamic_map.upper_bound(name);
 
         for (auto iter = lbound; iter != ubound; iter++) {
-            auto const& v = iter->second;
-            if (v.second == t) {
+            auto const& v = iter->second;  // pair(value, entry)
+            auto const& val = v.first;
+            auto const& ent = v.second;
+            if (ent == t) {
+                if (_df) {
+                    basic_stream bs;
+                    bs.printf("evict entry[%zi] %s=%s", entry, name.c_str(), val.c_str());
+                    _df(&bs);
+                }
                 _dynamic_map.erase(iter);
                 break;
             }
@@ -175,6 +279,11 @@ return_t http_header_compression_session::evict() {
         _dropped++;
     }
 
+    return ret;
+}
+
+return_t http_header_compression_session::duplicate(const std::string& name, const std::string& value) {
+    return_t ret = errorcode_t::success;
     return ret;
 }